move tooltip and render on load
